pyproject.toml
[tool.poetry]
name = "bagg-analysis"
version = "0.1.0"
description = ""
authors = ["lakishadavid <ltdavid2@illinois.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.35.79"
psycopg2-binary = "^2.9.10"
cfn-lint = "^1.22.2"
botocore = "^1.35.85"
matplotlib = "^3.4"
requests = "^2.32.3"
lineage = "^4.4.1"
pyliftover = "^0.4.1"
tqdm = "^4.67.1"
pandas = "^2.2.3"
inputimeout = "^1.0.4"
seaborn = "^0.13.2"
ipython = "^8.31.0"
numpy = ">=1.23"
scipy = "^1.14.1"
pygraphviz = "^1.14"
plotly = "^5.24.1"
cdlib = "^0.4.0"
python-decouple = "^3.8"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
